exon_detection_project/requirements.txt
# requirements.txt
numpy>=1.19.5
pandas>=1.3.0
biopython>=1.79
scikit-learn>=0.24.2
tensorflow>=2.5.0
matplotlib>=3.4.2
seaborn>=0.11.1
requests>=2.25.1
tqdm>=4.61.1
